requirements.txt
urllib3<2
google.generativeai
python-dotenv